refactor: move errapi to api/ to serve for both rest and graph api, #24
